requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
asttokens==3.0.0
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.7
decorator==5.2.1
Django==5.2.5
django-allauth==65.4.0
django-cors-headers==4.8.0
django-extensions==4.1
django-solo==2.4.0
django-tinymce==4.1.0
executing==2.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.30.0
idna==3.10
jedi==0.19.2
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
parso==0.8.4
pillow==11.3.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
spotipy==2.25.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.10.0
